api-userstate: отправка/получение cartstate с тестом

+++ website/src/main/java/ru/teamtwo/website/controller/user/OrderController.java
@@ -26,7 +26,7 @@ public class OrderController {
     private OrderRepository repository;
 
     @GetMapping("{id}")
-    public OrderDto get(@PathVariable long id){
+    public OrderDto get(@PathVariable Integer id){
         log.debug("get: {}", id);
         return new OrderDto(repository.getById(id));
     }
