+++ website/src/main/java/ru/teamtwo/website/repository/user/OrderItemRepository.java
@@ -3,5 +3,5 @@ package ru.teamtwo.website.repository.user;
 import org.springframework.data.jpa.repository.JpaRepository;
 import ru.teamtwo.website.model.user.OrderItem;
 
-public interface OrderItemRepository extends JpaRepository<OrderItem, Long> {
+public interface OrderItemRepository extends JpaRepository<OrderItem, Integer> {
 }
